seperate setting page to two container and update materian-ui
