+++ client/package.json
@@ -22,7 +22,7 @@
     "geolocation": "^0.2.0",
     "happypack": "^3.0.2",
     "history": "^3.2.1",
-    "material-ui": "^0.16.7",
+    "material-ui": "^0.19.2",
     "moment": "2.16.0",
     "react": "^15.4.2",
     "react-addons-perf": "^15.4.2",
